+++ java/src/main/java/edu/unialfa/java/service/DisciplinaService.java
@@ -2,7 +2,10 @@ package edu.unialfa.java.service;
 
 import edu.unialfa.java.model.Disciplina;
 import edu.unialfa.java.repository.DisciplinaRepository;
+import edu.unialfa.java.repository.TurmaDisciplinaRepository;
+import jakarta.persistence.EntityNotFoundException;
 import lombok.RequiredArgsConstructor;
+import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
 import java.util.List;
@@ -13,6 +16,9 @@ import java.util.Optional;
 public class DisciplinaService {
 
     private final DisciplinaRepository disciplinaRepository;
+
+    @Autowired
+    private TurmaDisciplinaRepository turmaDisciplinaRepository;
 
     public List<Disciplina> listarTodas() {
         return disciplinaRepository.findAll();
@@ -27,6 +33,17 @@ public class DisciplinaService {
     }
 
     public void excluir(Long id) {
+        if (!disciplinaRepository.existsById(id)) {
+            throw new EntityNotFoundException("Disciplina não encontrada com o ID: " + id);
+        }
+
+        boolean disciplinaVinculadaEmTurmaDisciplina = turmaDisciplinaRepository.existsByDisciplinaId(id);
+
+        if (disciplinaVinculadaEmTurmaDisciplina) {
+            throw new IllegalStateException("Não é possível excluir a disciplina. Ela está vinculada a uma ou mais turmas.");
+        }
+
         disciplinaRepository.deleteById(id);
     }
+
 }
